correciones en el pdf y campos nulos

+++ app/Http/Controllers/Api/V1/AsesorController.php
@@ -11,6 +11,11 @@ use Illuminate\Support\Facades\DB;
 
 class AsesorController extends Controller
 {
+    public function indexPublic()
+    {
+        $datos = Asesor::where('publico',true)->get();
+        return response()->json($datos);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -35,28 +40,29 @@ class AsesorController extends Controller
         }
         $files = $request->file('foto');
 
+        $asesor = new Asesor;
+
         if ($request->hasFile('foto')) {
 
             $nombre = uniqid() . '.' . $files->getClientOriginalName();
             $path = $carpeta . '/' . $nombre;
             \Storage::disk('public')->put($path, \File::get($files));
-
-            $asesor = new Asesor;
-            $asesor->nombre = $request->nombre;
-            $asesor->apellidos = $request->apellidos;
-            $asesor->rfc = $request->rfc;
-            $asesor->direccion = $request->direccion;
-            $asesor->email = $request->email;
-            $asesor->celular = $request->celular;
             $asesor->foto = $nombre;
-            $asesor->contacto_emergencia = $request->contacto_emergencia;
-            $asesor->save();
-
-            return response()->json($asesor);
-
-        } else {
-            return "error";
         }
+
+        $asesor->nombre = $request->nombre;
+        $asesor->apellidos = $request->apellidos;
+        $asesor->rfc = $request->rfc;
+        $asesor->direccion = $request->direccion;
+        $asesor->email = $request->email;
+        $asesor->celular = $request->celular;
+        $asesor->contacto_emergencia = $request->contacto_emergencia;
+        $asesor->status = $request->status;
+        $asesor->publico = $request->publico;
+        $asesor->save();
+
+        return response()->json($asesor);
+
 
 
     }
@@ -89,6 +95,8 @@ class AsesorController extends Controller
         $asesor->email = $request->email;
         $asesor->celular = $request->celular;
         $asesor->contacto_emergencia = $request->contacto_emergencia;
+        $asesor->status = $request->status;
+        $asesor->publico = $request->publico;
         $asesor->save();
         return response()->json($asesor);
     }
@@ -102,13 +110,17 @@ class AsesorController extends Controller
             return response()->json(['message' => 'Registro no encontrado'], 404);
         }
 
-        $nombreArchivo = $asesor->foto;
-        $files = $request->file('foto');
+        if ($request->hasFile('foto')) {
+            $nombreArchivo = $asesor->foto;
+            $files = $request->file('foto');
 
-        \Storage::disk('public')->delete($carpeta . '/' . $nombreArchivo);
-        $nombreNuevo = uniqid() . '.' . $files->getClientOriginalName();
-        $pathNuevo = $carpeta . '/' . $nombreNuevo;
-        \Storage::disk('public')->put($pathNuevo, \File::get($files));
+            \Storage::disk('public')->delete($carpeta . '/' . $nombreArchivo);
+            $nombreNuevo = uniqid() . '.' . $files->getClientOriginalName();
+            $pathNuevo = $carpeta . '/' . $nombreNuevo;
+            \Storage::disk('public')->put($pathNuevo, \File::get($files));
+            $asesor->foto = $nombreNuevo;
+
+        }
 
         $asesor->nombre = $request->nombre;
         $asesor->apellidos = $request->apellidos;
@@ -116,8 +128,9 @@ class AsesorController extends Controller
         $asesor->direccion = $request->direccion;
         $asesor->email = $request->email;
         $asesor->celular = $request->celular;
-        $asesor->foto = $nombreNuevo;
         $asesor->contacto_emergencia = $request->contacto_emergencia;
+        $asesor->status = $request->status;
+        $asesor->publico = $request->publico;
         $asesor->save();
         return response()->json($asesor);
     }
